Fix: 오늘의 복습문제 갯수 조회 로직 수정(#35) 복습문제 테이블과 오늘의 복습리스트 테이블 조인-> 매주복습량 테이블에서 가져오기

+++ back/src/main/java/com/clover/recode/domain/statistics/service/StatisticsService.java
@@ -9,7 +9,7 @@ public interface StatisticsService {
 
     StatisticsListRes getStatisticsList(Long id);
 
-    Long getReviewCnt(Long id);
+    Integer getReviewCnt(Long id);
 
      List<TodayProblemRes> getReviews(Long id);
